requirements.txt
asgiref==3.6.0
Django==4.2.1
django-multiselectfield==0.1.12
mypy==1.3.0
mypy-extensions==1.0.0
python-dotenv==1.0.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
